Move transaction to bin/glusterd2/transaction
