+++ bin/glusterd2/transaction/lock.go
