+++ bin/glusterd2/transaction/registry.go
