+++ commands/volumes/volume-create_test.go
@@ -6,12 +6,12 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/peer"
 	"github.com/gluster/glusterd2/pkg/api"
 	gderrors "github.com/gluster/glusterd2/pkg/errors"
 	"github.com/gluster/glusterd2/tests"
-	"github.com/gluster/glusterd2/transaction"
 	"github.com/gluster/glusterd2/volume"
 
 	"github.com/pborman/uuid"
