+++ commands/volumes/volume-start.go
@@ -4,9 +4,9 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/pkg/errors"
 	restutils "github.com/gluster/glusterd2/servers/rest/utils"
-	"github.com/gluster/glusterd2/transaction"
 	"github.com/gluster/glusterd2/volume"
 
 	"github.com/gorilla/mux"
